+++ konva.js
@@ -2111,6 +2111,12 @@
       this.setAttr('shadowBlur', blur * ratio * Math.min(scaleX, scaleY));
       this.setAttr('shadowOffsetX', offset.x * scaleX);
       this.setAttr('shadowOffsetY', offset.y * scaleY);
+    },
+    _applyGlobalCompositeOperation: function(shape) {
+      var globalCompositeOperation = shape.getGlobalCompositeOperation();
+      if (globalCompositeOperation !== 'source-over') {
+        this.setAttr('globalCompositeOperation', globalCompositeOperation);
+      }
     }
   };
   Konva.Util.extend(Konva.SceneContext, Konva.Context);
@@ -8421,8 +8427,10 @@
         var ratio = bufferCanvas.pixelRatio;
         if (hasShadow && !canvas.hitCanvas) {
           context.save();
+
           context._applyShadow(this);
           context._applyOpacity(this);
+          context._applyGlobalCompositeOperation(this);
           context.drawImage(
             bufferCanvas._canvas,
             0,
@@ -8433,6 +8441,7 @@
           context.restore();
         } else {
           context._applyOpacity(this);
+          context._applyGlobalCompositeOperation(this);
           context.drawImage(
             bufferCanvas._canvas,
             0,
@@ -8459,8 +8468,10 @@
           // apply shadow
           if (!caching) {
             context._applyOpacity(this);
+            context._applyGlobalCompositeOperation(this);
           }
           context._applyShadow(this);
+
           drawFunc.call(this, context);
           context.restore();
           // if shape has stroke we need to redraw shape
@@ -8473,6 +8484,7 @@
           context.save();
           if (!caching) {
             context._applyOpacity(this);
+            context._applyGlobalCompositeOperation(this);
           }
           context._applyShadow(this);
           drawFunc.call(this, context);
@@ -8480,6 +8492,7 @@
         } else {
           if (!caching) {
             context._applyOpacity(this);
+            context._applyGlobalCompositeOperation(this);
           }
           drawFunc.call(this, context);
         }
@@ -8876,6 +8889,27 @@
     1,
     Konva.Validators.alphaComponent
   );
+
+  Konva.Factory.addGetterSetter(
+    Konva.Shape,
+    'globalCompositeOperation',
+    'source-over'
+  );
+
+  /**
+     * get/set globalCompositeOperation of a shape
+     * @name globalCompositeOperation
+     * @method
+     * @memberof Konva.Shape.prototype
+     * @param {Number} blur
+     * @returns {Number}
+     * @example
+     * // get shadow blur
+     * var globalCompositeOperation = shape.globalCompositeOperation();
+     *
+     * // set shadow blur
+     * shape.globalCompositeOperation('source-in');
+     */
 
   Konva.Factory.addGetterSetter(Konva.Shape, 'shadowBlur');